LayoutTests/fast/webgpu/nocrash/fuzz-295370-expected.txt

CONSOLE MESSAGE: promises created
CONSOLE MESSAGE: Pass
